Factor out a function to write feed objects to a zip
This function assumes the input feed is well-formed. As the name
indicates, there are no sanity checks.

At Remix we sometimes need to transform GTFS files for downstream
systems that have particular requirements. This function makes
it a breeze to do those transformations with pandas and then
write out a zip file.